ai_service/requirements.txt
# DevMind AI Service - Python Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.7.0
chromadb>=0.5.3
python-dotenv>=1.0.0
httpx>=0.26.0
groq>=0.4.0
